fix(printStockInfo): 보유 상품 출력 위치 수정
새로 다른 상품을 구매할 때에도 재고 정보를 보여줘야 하므로 위치를 PurchaseController로 수정한다

추후 SupplyController의 존재 의미에 대해 고민한다
